+++ src/Core/Dfe.Complete.Application/Common/Models/TrustDetailsDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Dfe.Complete.Domain.Entities
+﻿namespace Dfe.Complete.Application.Common.Models
 {
 	public class TrustDetailsDto
 	{
